Move custom payload conversion to separate file
